ADSPRPC: allow utilization of SMD instead of GLINK
Legacy platforms require SMD, and do not support GLINK.
Since the default should be GLINK moving forward, then add SMD support as a special case.
To do that, the DT property qcom,fastrpc-smd is added, along with its handling logic.
This replaces the glink flag, and inverts the logic to always use GLINK calls as default.

Change-Id: Ie71fc6fcc6d389127b47efd8790bc61241497728