drivers: input: touchscreen: locally decide values of reg_offsets
registry offsets have fixed values per touchscreen model.
Hence, there is no need to declare them in dt then reference them.
Simply declare the values in the driver when initializing the driver.
The same logic used to reference values from dt is used to decide the values in the driver.
I.e, a combination of chip_id and firmware_revision_extra decide which set of offsets to use.

Change-Id: I4175f94e5fa9ba82b65594d184c6fa6d1e279815