input: misc: fpc1145: Use TRIGGER_HIGH for FPC interrupt
Using TRIGGER_RISING we are risking to fire the interrupt
handler way too many times, even when we don't really
need it to.
Actually, we want it to fire when the interrupt is HIGH,
because our actions will be taken ONLY in this case:
it'd be useless to read the GPIO state in the handler
when the IRQ is rising when we can just configure it as
trigger high.

Change-Id: Ic9038615ca3beecb27518c2d90eea4e21ad18031