input: misc: fpc1145: Set a wakelock on IRQ IOCTL for TZ processing
When userspace calls this IOCTL, it's waiting for FINGER DOWN
event. Once this event happens, the userspace will send other
commands to the TZ app to either retry or start finger image
acquire.

In that case we must ensure that the kernel is alive and that
it will not try to suspend, otherwise a TZ fault can happen,
leading to machine reboot.

Change-Id: I7e6862ae4debc39855655f2de8d502ce58bf6281