fbdev: msm: somc_panel: Use B sequence as poweroff reset sequence
The stock panel driver uses the B sequence as a normal poweroff reset sequence, if found.
https://github.com/sonyxperiadev/kernel-copyleft/blob/41.2.A.0.xxx/drivers/video/msm/mdss/mdss_dsi_panel_driver.c#L1109-L1121
This is not the case in our driver, and leads to issues with dt2w on hybrid incell panels.
The reason for that is that the devicetree nodes of hybrid incell panels only declare the B sequences (somc,pw-off-rst-b-seq).
The missing normal poweroff reset sequence causes the driver to always report a fatal error upon turning off the panel:
[  967.464295] FATAL: poweroff reset sequence is NULL!!

The solution is to add a check to make sure that there's always a poweroff reset sequence set, by falling back to the B
sequence. This makes the driver behaviour similar to that on stock.

Change-Id: I69cf18e4ba3fc2081e89e37e4eb6d6c86299ca48